+++ spec/immer/storage.spec.ts
@@ -1,9 +1,9 @@
 import { timer } from 'rxjs';
 import { tap, map } from 'rxjs/operators';
-import { StateKeeper } from '../../src/ng-state/state/history';
-import { Store } from '../../src/ng-state/store/store';
-import { NgStateTestBed } from '../../src/ng-state/ng-state.test-bed';
-import { PersistStateStorage, PersistStateManager } from '../../src/ng-state/store/plugins/persist-state.plugin';
+import { StateKeeper } from '@ng-state/store';
+import { Store } from '@ng-state/store';
+import { NgStateTestBed } from '@ng-state/store';
+import { PersistStateStorage, PersistStateManager } from '@ng-state/store';
 import { ImmerDataStrategy } from '@ng-state/immer-data-strategy';
 
 jest.useFakeTimers();