+++ src/app/immutable-app/actions/todo.actions.ts
@@ -1,6 +1,6 @@
 import * as Immutable from 'immutable';
 
-import { HasStore, InjectStore } from '../../../ng-state/decorators/inject-store.decorator';
+import { HasStore, InjectStore } from '@ng-state/store';
 
 @InjectStore(['${stateIndex}'])
 export class TodoStateActions extends HasStore<Immutable.Map<any, any>> {