+++ src/app/immutable-app/actions/todos.actions.ts
@@ -1,5 +1,5 @@
 import * as Immutable from 'immutable';
-import { HasStore, InjectStore } from '../../../ng-state/decorators/inject-store.decorator';
+import { HasStore, InjectStore } from '@ng-state/store';
 import { TodoModel } from './todo.model';
 
 @InjectStore('todos')